requirements.txt
groq
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
gspread
pandas
openpyxl
gspread-formatting